Deprecate legacy code
